Base changes
Connected the aqua farm with our underground storage tunnel and
attempted a rail network between them.
Added more fridge storage to the kitchen, moved dry goods to the sink
container, and organized food so that ingredients are in the big fridges
while the small fridges contain prepared foods. Added things to the
armory area and reorganized containers and got rid of junk weapons that
were obsolete. Made some cosmetic changes to the base in various areas.